Revert "Avoid eager torchaudio imports"
This reverts commit ed1c2724f570798aa2785109c1f095640667c4fb.
